feat(contracts): overhaul from priority-list data structure case

+++ contracts/interfaces/IERC20Minimal.sol
@@ -1,4 +1,4 @@
-// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: BUSL-1.1
 
 // OpenZeppelin Contracts (last updated v4.6.0) (token/ERC20/IERC20.sol)
 
@@ -8,5 +8,5 @@ pragma solidity ^0.8.10;
  * @dev Interface of the ERC20 standard as defined in the EIP.
  */
 interface IERC20Minimal {
-    function decimals() external view returns (uint256);
+    function decimals() external view returns (uint8);
 }
